Changed the way buildings are added when a city grows.
Removed the nested while(true) loops because it's bad practice.
New houses also have rotations in cardinal directions instead of facing towards the center.
Bug fixed where unity would spit out an outOfBoundsException because the gridTiles list was empty and we tried to access a value in the list anyway